+++ libs/frontend/modules/builder/src/renderer/atoms/index.ts
@@ -1,0 +1,1 @@
+export * from './atomFactory'
