+++ libs/frontend/modules/builder/src/renderer/index.ts
@@ -1,6 +1,3 @@
-export * from './atoms/atomFactory'
-export * from './defaultRenderContext'
-export * from './querySelectorRenderedElement'
-export * from './reactComponentFactory'
+export * from './atoms'
 export * from './Renderer'
-export * from './renderFactory'
+export * from './utils'
