+++ libs/frontend/modules/builder/src/renderer/pipes/renderAtomPipe.tsx
@@ -1,9 +1,9 @@
 import { mergeProps } from '@codelab/shared/utils'
 import { css } from '@emotion/react'
 import React from 'react'
-import { reactComponentFactory } from '../reactComponentFactory'
 import { evalCss } from '../utils/evalCss'
 import { onRendered } from '../utils/onRendered'
+import { reactComponentFactory } from '../utils/reactComponentFactory'
 import { RenderPipeFactory } from './types'
 
 /**
