+++ libs/frontend/modules/builder/src/renderer/pipes/renderPipeline.tsx
@@ -1,19 +1,19 @@
 import { compose } from 'ramda'
-import { basePropsPipe } from './pipes/basePropsPipe'
-import { componentPipe } from './pipes/componentPipe'
-import { conditionalRenderPipe } from './pipes/conditionalRenderPipe'
-import { extraElementPropsPipe } from './pipes/extraElementPropsPipe'
-import { extraPropsPipe } from './pipes/extraPropsPipe'
-import { hookPipe } from './pipes/hookPipe'
-import { loopingRenderPipe } from './pipes/loopRenderingPipe'
-import { persistedPropsPipe } from './pipes/persistedPropsPipe'
-import { propMapBindingsPipe } from './pipes/propMapBindingsPipe'
-import { propsWithTypePipe } from './pipes/propsWithTypePipe'
-import { propTransformationJsPipe } from './pipes/propTransformationJsPipe'
-import { reactNodePipe } from './pipes/reactNodePipe'
-import { renderAtomPipe } from './pipes/renderAtomPipe'
-import { renderChildrenPipe } from './pipes/renderChildrenPipe'
-import { renderPropsPipe } from './pipes/renderPropsPipe'
+import { basePropsPipe } from './basePropsPipe'
+import { componentPipe } from './componentPipe'
+import { conditionalRenderPipe } from './conditionalRenderPipe'
+import { extraElementPropsPipe } from './extraElementPropsPipe'
+import { extraPropsPipe } from './extraPropsPipe'
+import { hookPipe } from './hookPipe'
+import { loopingRenderPipe } from './loopRenderingPipe'
+import { persistedPropsPipe } from './persistedPropsPipe'
+import { propMapBindingsPipe } from './propMapBindingsPipe'
+import { propsWithTypePipe } from './propsWithTypePipe'
+import { propTransformationJsPipe } from './propTransformationJsPipe'
+import { reactNodePipe } from './reactNodePipe'
+import { renderAtomPipe } from './renderAtomPipe'
+import { renderChildrenPipe } from './renderChildrenPipe'
+import { renderPropsPipe } from './renderPropsPipe'
 
 //
 // Construct the pipeline:
