+++ libs/frontend/modules/builder/src/renderer/utils/defaultRenderContext.ts
@@ -1,5 +1,5 @@
-import { renderPipeline } from './renderPipeline'
-import { RenderContext } from './types/RenderTypes'
+import { RenderContext } from '../pipes'
+import { renderPipeline } from '../pipes/renderPipeline'
 
 export const defaultRenderContext = (
   context: Omit<RenderContext, 'render'>,
