+++ libs/frontend/modules/builder/src/renderer/utils/querySelectorRenderedElement.ts
